# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InstaClone API",
        "description": "API REST para rede social inspirada no Instagram.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cadastrar usuário",
                "operationId": "b22b8f3540bd56ac6281fb4bedcd4f6f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "username",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Matheus"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "matheus"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "matheus@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "senha1234"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "senha1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "0af001e4a6a482c7e4c55b55078cba05",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "matheus@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "senha1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login efetuado"
                    },
                    "422": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "30b5cbc0a8f47628a19c9c7f67cc04b0",
                "responses": {
                    "200": {
                        "description": "Desconectado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Renovar token",
                "operationId": "0c73fc1dcf46e4858be17f4357754338",
                "responses": {
                    "200": {
                        "description": "Novo token gerado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Usuário autenticado",
                "operationId": "fcbd71bab9d44aa1052f12b598e9019e",
                "responses": {
                    "200": {
                        "description": "Dados do usuário"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{username}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Perfil por username",
                "operationId": "65b56cf32dfdab89290fbdaa7ffb1c07",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Perfil do usuário"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/me": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Editar perfil",
                "operationId": "237401c64b022e14a457c9eb0c82ac3e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "bio": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil atualizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/me/avatar": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Atualizar avatar",
                "operationId": "29ab710de8a061c5caa5a196f8a3b562",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Avatar atualizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/search": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Buscar usuários",
                "operationId": "fe6369254bd20cc5144ec156a52beec4",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de usuários"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/suggestions": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Sugestões de quem seguir",
                "operationId": "4a79b2cb08d31bcab48b787cd308546f",
                "responses": {
                    "200": {
                        "description": "Lista de sugestões"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}/posts": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Posts do usuário",
                "operationId": "df8eccb0395d72e50f6085314c90ccbe",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de posts"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}/follow": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Seguir usuário",
                "operationId": "88a0c8506619e6d542cd5630701298ea",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Seguindo"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Deixar de seguir",
                "operationId": "a9071961a36adc35ddc7532c3775f29f",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deixou de seguir"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}/followers": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Seguidores",
                "operationId": "1869084cfd26a9c23a6f47761d7e9262",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de seguidores"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}/following": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Quem o usuário segue",
                "operationId": "3399782c2987809709ccbd36e73fad6d",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de seguidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}/is-following": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Verificar se segue",
                "operationId": "dabc50db731e98194779ae7eb138343b",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status de follow"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/feed": {
            "get": {
                "tags": [
                    "Feed"
                ],
                "summary": "Feed paginado por cursor",
                "operationId": "6bfa6990f1396df28ea356171d44674f",
                "parameters": [
                    {
                        "name": "cursor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Posts do feed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Criar post",
                "operationId": "cb273c39513bc7595806b73072fea0f7",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "caption": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post criado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/posts/{post}/likers": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Quem curtiu",
                "operationId": "7e92c5f4d661532399c6902a56345e8e",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de usuários que curtiram"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}